move tests/internal -> tests/tracer
