+++ tests/tracer/runtime/test_metric_collectors.py
@@ -9,7 +9,7 @@ from ddtrace.internal.runtime.constants import (
     GC_RUNTIME_METRICS,
     PSUTIL_RUNTIME_METRICS,
 )
-from ... import BaseTestCase
+from tests import BaseTestCase
 
 
 class TestRuntimeMetricCollector(BaseTestCase):
