+++ tests/tracer/runtime/test_metrics.py
@@ -1,7 +1,7 @@
 import mock
 from ddtrace.internal.runtime.collector import ValueCollector
 
-from ... import BaseTestCase
+from tests import BaseTestCase
 
 
 def mocked_collector(mock_collect, **kwargs):
