+++ tests/tracer/test_context_manager.py
@@ -4,7 +4,7 @@ from ddtrace.context import Context
 from ddtrace.internal.context_manager import DefaultContextManager
 from ddtrace.span import Span
 
-from .. import BaseTestCase
+from tests import BaseTestCase
 
 
 class TestDefaultContextManager(BaseTestCase):
