+++ tests/tracer/test_import_hooks.py
@@ -1,16 +1,60 @@
 import mock
-
 import pytest
 
-from ddtrace.internal.import_hooks import ModuleHookRegistry
-from ddtrace.internal.import_hooks import hooks as global_hooks
+import tests.internal
+import tests.test_module
 
-from tests.subprocesstest import run_in_subprocess, SubprocessTestCase
+from ddtrace.internal.import_hooks import register_module_hook, ModuleHookRegistry, hooks as global_hooks
+from tests import SubprocessTestCase
+from tests.subprocesstest import run_in_subprocess
 
 
 @pytest.fixture
 def hooks():
-    return ModuleHookRegistry()
+    yield ModuleHookRegistry()
+
+
+def test_register_module_hook_decorator(hooks):
+    """
+    When registering a module hook function
+        Using a decorator to register the hook
+            We properly register the hook
+    """
+    # Create and register a module hook
+    module_hook = mock.Mock()
+
+    # DEV: This is the same as:
+    #   @register_module_hook('module')
+    #   def module_hook(module):
+    #       pass
+    hook_res = register_module_hook("module", registry=hooks)(module_hook)
+
+    # Ensure the returned value is our original function
+    assert hook_res == module_hook
+
+    # Ensure the hook was registered
+    module = object()
+    hooks.call("module", module=module)
+    module_hook.assert_called_once_with(module)
+
+
+def test_register_module_hook_func(hooks):
+    """
+    When registering a module hook function
+        Passing the hook as a parameter
+            We properly register the hook
+    """
+    # Create and register a module hook
+    module_hook = mock.Mock()
+    hook_res = register_module_hook("module", module_hook, registry=hooks)
+
+    # Ensure the returned value is our original function
+    assert hook_res == module_hook
+
+    # Ensure the hook was registered
+    module = object()
+    hooks.call("module", module=module)
+    module_hook.assert_called_once_with(module)
 
 
 @pytest.fixture
@@ -28,7 +72,7 @@ def test_global_hooks():
     assert isinstance(global_hooks, ModuleHookRegistry)
 
     # No hooks are registered by default
-    assert len(global_hooks.hooks) == 0
+    assert len(tests.tracer.test_import_hooks.hooks) == 0
 
 
 def test_registry_init(hooks):
@@ -36,7 +80,7 @@ def test_registry_init(hooks):
     When initializing a new registry hook
         No hooks are added by default
     """
-    assert len(hooks.hooks) == 0
+    assert len(tests.tracer.test_import_hooks.hooks) == 0
 
 
 def test_registry_reset(hooks, module_hook):
@@ -48,13 +92,13 @@ def test_registry_reset(hooks, module_hook):
     hooks.register("test.module.name", module_hook)
     hooks.register("pytest", module_hook)
     hooks.register("ddtrace", module_hook)
-    assert len(hooks.hooks) == 3
+    assert len(tests.tracer.test_import_hooks.hooks) == 3
 
     # Reset the registry
     hooks.reset()
 
     # All hooks are removed
-    assert len(hooks.hooks) == 0
+    assert len(tests.tracer.test_import_hooks.hooks) == 0
 
 
 def test_registry_call_with_module(hooks):
@@ -156,7 +200,7 @@ def test_registry_call_no_name(hooks):
     module_name = "test.module.name"
 
     # Ensure the module isn't registered
-    assert module_name not in hooks.hooks
+    assert module_name not in tests.tracer.test_import_hooks.hooks
 
     # Call the hooks, this should not have any side effects
     hooks.call(module_name, module)
