+++ tests/tracer/test_import_hooks_integration.py
