+++ tox.ini
@@ -23,7 +23,6 @@ envlist =
     {py27,py35,py36,py37,py38}-tracer
     {py27,py35,py36,py37,py38,py39}-profile{,-gevent}
     {py27,py35,py36,py37,py38,py39}-profile-minreqs{,-gevent}
-    {py27,py35,py36,py37,py38}-internal
     {py27,py35,py36,py37,py38}-integration
     {py27,py35,py36,py37,py38}-ddtracerun
     {py27,py35,py36,py37,py38}-test_utils
@@ -466,8 +465,6 @@ commands =
 # run only essential tests related to the tracing client
     tracer: pytest {posargs} tests/tracer
     profile: python -m tests.profiling.run pytest --capture=no --verbose {posargs} tests/profiling
-# run only the `ddtrace.internal` tests
-    internal: pytest {posargs} tests/internal
 # run only the opentrace tests
     opentracer: pytest {posargs} tests/opentracer/core
     opentracer_asyncio: pytest {posargs} tests/opentracer/test_tracer_asyncio.py
